build: no longer depend on gulp for lint tasks
* Fixes the `prefer-literal` lint rule not actually running because it wasn't loaded in the tslint.json.
* Converts the CODEOWNERS linting task into a regular Node script that doesn't depend on Gulp.
* Turns the `stylelint` and `tslint` tasks into npm scripts.
* Runs the new npm scripts directly rather than through Gulp on the CI.
