+++ tools/gulp/tasks/ci.ts
@@ -1,7 +1,5 @@
 import {task} from 'gulp';
 
-task('ci:lint', ['lint']);
-
 // Gulp sometimes does not exit properly on CI. This is to prevent that.
 // TODO(devversion): look if there is some blocking child process.
 task('ci:test', ['test:single-run'], () => process.exit(0));
